trigger-http: Move headers tests into headers mod
